Exempi.parse_bitmask: long names by default
Most of exempi is a loose enough wrapper that all of the option
names use the verbose, C-like names. .parse_bitmask, however, was
returning shorthand option names that are more akin to how
Fasttrack deals with them. Obviously that was inappropriate, so now
the long form is returned by default with an option to return
short versions instead.